DefaultTrace -> ObservedTrace, EmptyTrace -> UnobservedTrace
